refactor: api & queries 폴더 정리
