+++ src/api/gatherings-workspace/gatherings-workspace.ts
@@ -1,32 +1,17 @@
-import { clientAxios, serverAxios } from "@/lib/axios";
+import { clientAxios } from "@/lib/axios";
 import { GatheringCreateFormData } from "@/types/category";
+import { Pagination } from "@/types/pagination";
 
+// 모임 조회
+export const gatheringGetApi = async (page: Pagination) => {
+  const { data } = await clientAxios.get(`api/gatherings/workspace?offset=${page.offset}&limit=${page.limit}`);
+  return data.data;
+};
+
+// 모임 생성
 export const gatheringCreateApi = async (formData: GatheringCreateFormData) => {
   const { data } = await clientAxios.post("/api/gatherings/workspace", formData);
   return data;
-};
-
-export const gatheringDetailGetApi = async (id: number) => {
-  const { data } = await serverAxios.get(`/api/gatherings/${id}`);
-  return data.data;
-};
-
-// 모임 상세 클라이언트 데이터
-export const gatheringDetailGetApiClient = async ({ queryKey }: any) => {
-  const { data } = await clientAxios.get(`/api/gatherings/${queryKey[1]}`);
-  return data.data;
-};
-
-// 모임 참여
-export const gatheringJoinApi = async (id: number) => {
-  await clientAxios.post(`/api/gatherings/join/${id}`);
-  return id;
-};
-
-// 모임 취소
-export const gatheringJoinCancelApi = async (id: number) => {
-  await clientAxios.delete(`/api/gatherings/join/${id}`);
-  return id;
 };
 
 // 모임 수정
