+++ src/queries/gatherings-workspace/useGatheringDelete.ts
@@ -1,4 +1,4 @@
-import { gatheringDeleteApi } from "@/api/gatherings";
+import { gatheringDeleteApi } from "@/api/gatherings-workspace/gatherings-workspace";
 import { toast } from "@/hooks/use-toast";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
